Moved pages to the corresponding folder
