+++ automation_testing/pages/DataView/PaginationComponent.ts
@@ -1,4 +1,4 @@
-import { BasePage } from "../pages/BasePage";
+import { BasePage } from "../BasePage";
 import { Locator, Page } from "@playwright/test";
 
 export class PaginationComponent extends BasePage {
