+++ automation_testing/tests/dataview/filter.spec.ts
@@ -1,6 +1,6 @@
 import { expect, test } from "@playwright/test";
-import { DataviewPage } from "../../pages/DataViewPage";
-import { FilterComponent } from "../../pages/FilterComponent";
+import { DataviewPage } from "../../pages/DataView/DataViewPage";
+import { FilterComponent } from "../../pages/DataView/FilterComponent";
 import { dataview_data, filter_data } from "../../utils/data/dataview_data";
 import { addQuotes, addComma } from "../../utils/helpers/helper";
 test.describe("DataView - Filter", () => {
